python/day_07/challenge_03/api-project/pyproject.toml
[tool.poetry]
name = "api-project"
version = "0.1.0"
description = ""
authors = ["W1NDCOR3 <diogo.romero.paz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
